mapmetadata: Map health metadata onto predefined research domains

Prior to gaining full access to health datasets, explore publicly available metadata and map metadata onto predefined research domains. This package uses structural metadata files downloaded from the Health Data Research Gateway (https://healthdatagateway.org/en). In theory, any metadata file with the same structure as the files downloaded from this gateway can be used with this package, but the package has been developed and tested on metadata files from this gateway only.

See also

Author

Maintainer: Rachael Stickland rstickland@turing.ac.uk (ORCID)

Other contributors:

Batool Almarzouq (ORCID) [contributor]

Mahwish Mohammad (ORCID) [contributor]

Daniel Delbarre (ORCID) [contributor]

Nida Ziauddeen (ORCID) [contributor]

Zoë Turner (Zoë reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/674) [reviewer]

Yohann Mansiaux (Yohann reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/674) [reviewer]